change login to js
